Switch to a regular function for the calculation.
This is much easier to understand than a class method that uses mutable
class variables. Now we don't have to worry about setters and their side
effects, we just need to pass in some Options directly.

Added support for parallelization during the multiplication. Also return
a dedicated Result class that allows more diagnostics in the future.